refactor to base consent store model
